Fix Python 3 division yields a float when an int is expected
A handful of tests failed because there was a division by 2 where the Python 2 behavior was expected. Python 3 division always yields a float. Added a cast to int to restore old behavior.

+++ cnfformula/cnf.py
@@ -1029,7 +1029,7 @@ class CNF(object):
         -------
             a list of clauses
         """
-        threshold = (len(variables)+1)/2
+        threshold = int((len(variables)+1)/2)
         return cls.greater_or_equal_constraint(variables, threshold)
 
     @classmethod
@@ -1045,7 +1045,7 @@ class CNF(object):
         -------
             a list of clauses
         """
-        threshold = len(variables)/2
+        threshold = int(len(variables)/2)
         return cls.less_or_equal_constraint(variables, threshold)
      
     @classmethod
@@ -1061,7 +1061,7 @@ class CNF(object):
         -------
             a list of clauses
         """
-        threshold = (len(variables)+1)/2
+        threshold = int((len(variables)+1)/2)
         return cls.equal_to_constraint(variables,threshold)
      
     @classmethod
@@ -1077,7 +1077,7 @@ class CNF(object):
         -------
             a list of clauses
         """
-        threshold = len(variables)/2
+        threshold = int(len(variables)/2)
         return cls.equal_to_constraint(variables,threshold)
 
     class unary_mapping(object):
